fix: STPM MUET float support + i18n for IC/verify-email pages
- Change min_muet_band from IntegerField to FloatField (65 courses have
  fractional bands like 2.5, 3.5, 4.5)
- Fix parser MUET fallback for courses with MUET in <div> not <li>
- Fix fixture converter int() truncation of fractional MUET values
- Regenerate stpm_requirements fixture with corrected MUET values
- Sync all 1,112 rows to Supabase (column + data)
- STPM golden master rebaselined: 1995 → 1976
- Add i18n translation keys for IC claim and verify-email pages

+++ halatuju_api/apps/courses/models.py
@@ -756,7 +756,7 @@ class StpmRequirement(models.Model):
     spm_subject_group = models.JSONField(null=True, blank=True)
 
     # MUET requirement
-    min_muet_band = models.IntegerField(default=1)
+    min_muet_band = models.FloatField(default=1.0)
 
     # Demographic / fitness requirements
     req_interview = models.BooleanField(default=False)
